Renamed export files to index to decrease confusion
