+++ Controllers/Applicants.ts
@@ -5,8 +5,8 @@ import {
   CreateApplicantResponseInput,
 } from "../types/main";
 import sendEmail from "../utils/sendEmail";
-import * as Openings from "../models/Openings/Openings";
-import * as Applicants from "../models/Applicants/Applicants";
+import * as Openings from "../models/Openings/index";
+import * as Applicants from "../models/Applicants/index";
 import _ from "lodash";
 import Joi from "joi";
