+++ models/Applicants/deleteApplicant.ts
@@ -5,7 +5,7 @@ import {
 import { Dynamo } from "../../awsClients/ddbDocClient";
 import { ENTITY_TYPES } from "../../Config";
 import { DeleteApplicantInput } from "../../types/main";
-import { getApplicantById } from "./Applicants";
+import { getApplicantById } from ".";
 const { DYNAMO_TABLE_NAME } = process.env;
 
 export default async function Remove(
